+++ CHANGELOG.rst
@@ -4,9 +4,17 @@ Changelog
 3.0.0b12 (unreleased)
 +++++++++++++++++++++
 
+Features:
+
+- The behavior to apply when encountering unknown fields while deserializing can be controlled with the ``unknown`` option (:issue:`524`).
+  It makes it possible to either "include", "exclude" or "raise".
+
+Other changes:
+
 - *Backwards-incompatible*: Pre/Post-processors MUST return modified data.
   Returning ``None`` does not imply data were mutated (:issue:`347`). Thanks
   :user:`tdevelioglu` for reporting.
+
 
 3.0.0b11 (2018-05-20)
 +++++++++++++++++++++